+++ translations/holyscreen_en_US.ts
@@ -1079,6 +1079,11 @@
         <translation>Search media</translation>
     </message>
     <message>
+        <location filename="../src/ui/operator/LibraryPanel.qml" line="183"/>
+        <source>Confirmar busca</source>
+        <translation>Confirm search</translation>
+    </message>
+    <message>
         <location filename="../src/ui/operator/LibraryPanel.qml" line="173"/>
         <source>Letras</source>
         <translation>Lyrics</translation>